+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @elf-framework/ui
 
+## 0.0.84
+
+### Patch Changes
+
+- fixed delay time in toast.hide()
+
+## 0.0.83
+
+### Patch Changes
+
+- fixed zIndex to z-index in View component
+
+## 0.0.81
+
+### Patch Changes
+
+- update alert, toast props
+
 ## 0.0.79
 
 ### Patch Changes